Add terrors.IsRetryable() helper
Add a package-level version of Error.IsRetryable(error) which interally
performs the necessary cast, same as terrors.PrefixMatches() etc.

This is intended to make it easier for users to write handler functions
that return differently depending whether the error is retryable or not,
when the "retryableness" of the top-level error is not immediately
obvious because it could have originated from several different places
inside the handler.