Clarify when `Is` behaves differently to `PrefixMatches`
Based on a quick read of the function definition/docs, I had
assumed that `Is` was used to get errors in a stack created by
calls to `Augment`.

This is not the case since `Augment` preserves duplicates the
code from the wrapped error to the wrapping error.

It is only when used in conjunction with `NewInternalWithCause`
that `Is` is useful.