Added terrors.Verbose()
This allows you to mark terror.Errors as "verbose". The most useful
upshot of this, as mentioned in the godoc, is so that you can get more
information in a failed test if you write

```go
require.NoError(t, terrors.Verbose(err))
```

This will cause `err.Error()` on the returned error to return
the `VerboseString()` rather than the regular short one.